+++ src/components/SearchModal/ListSelect.tsx
@@ -4,9 +4,10 @@ import ReactGA from 'react-ga'
 import { useDispatch, useSelector } from 'react-redux'
 import { Text } from 'rebass'
 import styled from 'styled-components'
+import { useFetchListCallback } from '../../hooks/useFetchListCallback'
 import { AppDispatch, AppState } from '../../state'
 import { acceptListUpdate, removeList, selectList } from '../../state/lists/actions'
-import { useFetchListCallback, useSelectedListUrl } from '../../state/lists/hooks'
+import { useSelectedListUrl } from '../../state/lists/hooks'
 import { CloseIcon, LinkStyledButton, TYPE } from '../../theme'
 import listVersionLabel from '../../utils/listVersionLabel'
 import { parseENSAddress } from '../../utils/parseENSAddress'
